+++ v2/interfaces/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.31.1 (2024-05-01)
+
+**Note:** Version bump only for package @chain-registry/interfaces
+
+
+
+
+
 # [0.31.0](https://github.com/cosmology-tech/chain-registry/compare/@chain-registry/interfaces@0.30.0...@chain-registry/interfaces@0.31.0) (2024-04-30)
 
 **Note:** Version bump only for package @chain-registry/interfaces